refactor MockSQLiteDB
