+++ tests/workflows/pipeline_modules/test_segmentation.py
@@ -26,7 +26,7 @@ from ophys_etl.workflows.pipeline_modules.segmentation import \
     SegmentationModule  # noqa E402
 from ophys_etl.workflows.well_known_file_types import WellKnownFileType # noqa E402
 from ophys_etl.workflows.workflow_steps import WorkflowStep # noqa E402
-from ophys_etl.test_utils.db_utils import MockSQLiteDB
+from ophys_etl.test_utils.db_base import MockSQLiteDB
 
 class TestSegmentation(MockSQLiteDB):
